requirements.txt
Flask==3.0.2
PyYAML==6.0.1
python-dateutil==2.9.0.post0
gpuinfo~=1.0.0a7
# bilibili-api-python==16.2.0
git+https://github.com/valkjsaaa/bilibili-api.git@531168565d50c7f6e7968d24147b10676e6f2c0b
Quart==0.19.4
srt==3.5.3
chardet==5.2.0
requests==2.31.0
rsa==4.9
toml~=0.10.2
urllib3==2.2.1
pillow==10.2.0